Updated `annotation#frontend` to create an accessor function for getting the property type.
This commit creates a function that is intended to be used for getting
the globally configured property type. The mentioned function is then
used any time the property name is needing to be fetched. The purpose of
this is so that the functions within `annotation#frontend` can be given
an explicit property type or fall-back to whatever is configured.

To avoid grabbing the property while the namespace is being compiled,
the default parameter was changed from `g:annotation_property` to
`v:null`, and the accessor was updated to check if its parameter is
`v:null` before trying to resolve `g:annotation#property`. This way, any
changes made to `g:annotation_property` by the user will be recognized
as the default parameter for the aforementioned functions.